+++ src/compiler/scanner/mod.rs
@@ -1,0 +1,6 @@
+mod lines_0_1000;
+pub use lines_0_1000::*;
+mod lines_1000_2000;
+pub use lines_1000_2000::*;
+mod lines_2000_end;
+pub use lines_2000_end::*;
